requirements.txt
Flask~=2.1.2
matplotlib~=3.5.2
pandas~=1.4.2
requests~=2.28.0
Pillow~=9.1.1
pytest~=7.1.2
pytest-cov~=3.0.0
flask-sqlalchemy~=2.5.1
xlwt~=1.3.0
openpyxl~=3.0.10
xlrd~=2.0.1
